Fix suggestions not appearing issue in windows
Completions are not working for the untitled files in composer in an windows environment. With this fix, introduce a new File pattern matching regex to avoid the issue.
Issue was due to the Lang server not being able to identify the untitled file reuest since the pattern had not addressed the proper windows file pattern.