Fix for parsing erroneous tokens for Completions
This fix make use of two extension points in the ANTLR default
ErrorHandler; reportMatch() and sync().

reportMatch() triggered per each successfull token match. When we are in
a token match this validates whether we're in the last terminating
token(i.e. '}','{',',',';').

Refer below example;

1. import ballerina/http;
2. import ballerina/io;
3. import ballerina/|

Suppose your cursor is at the end of the line #3. Your last terminating
token is the ';' in the line #2. If the current token of the parser is
above last terminating token; we'll remove/consume all future tokens
up to the current cursor position.

sync() triggered before any decision point in the parser. When we are in
a sync() hit; we'll analyze whether the cursor is just after the first
token of the cursor line. Whenever; sync hits before mentioned token it
will be removed/consumed by the CustomErrorStrategy.