Add code action for create variable
Whenever there's a function with a return type; and when we don' t
assign the return into a variable we get "variable assignment is required"
error message. It would be nice to have a code action to create a
variable for the return value of the function.

Thus, this PR resolves https://github.com/ballerina-platform/ballerina-lang/issues/10240